### 2025-08-28 #8881 @samhforbes
Hi @truenomad just looking at this error I think the paper can be in the main repository, or a branch of the main repository, but not in a different repository entirely as far as I know.
