### 2025-08-28 #8881 @truenomad
> Hi [@truenomad](https://github.com/truenomad) just looking at this error I think the paper can be in the main repository, or a branch of the main repository, but not in a different repository entirely as far as I know.

Ahh ok! Just managed to change location of paper now. Hopefully it satisfies the bots requirements. :D
