### 2025-08-29 #8881 @truenomad
@samhforbes I am realisng whats happening. The actual repository for the paper should be : https://github.com/truenomad/AgePopDenom_Paper_JOSS and not https://github.com/truenomad/AgePopDenom (this is the software). How do i change this ?
