### 2025-09-19 #8881 @samhforbes
Sorry I missed your update. The paper should be within the same repo as the code, though you can have it in a separate branch if you prefer. Please see our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss)
