### 2026-01-14 #8881 @truenomad
@arfon Thanks for the suggestions. Of the recommended papers, three (`APCtools`, `sptotal`, and `ungroup`) are directly relevant to our work, and we have now included and discussed them in the related software review section.
