### 2026-01-15 39e564f
Add JOSS paper
