+++ paper.md
@@ -1,0 +1,101 @@
+---
+title: "AgePopDenom: Geostatistical Modeling for Fine-Scale Population Age-Structures"
+tags:
+  - R
+  - geostatistics
+  - population modeling
+  - demographics
+  - spatial analysis
+  - public health
+authors:
+  - name: Mohamed A. Yusuf
+    orcid: 0000-0002-9339-4613
+    affiliation: "1, 2"
+    corresponding: true
+  - name: Victor A. Alegana
+    orcid: 0000-0001-5177-9227
+    affiliation: 1
+  - name: Chibuzor Christopher Nnanatu
+    orcid: 0000-0002-5841-3700
+    affiliation: "3, 4"
+  - name: Andrew Tatem
+    orcid: 0000-0002-7270-941X
+    affiliation: 3
+affiliations:
+  - name: The World Health Organization Regional Office for Africa (WHO AFRO)
+    index: 1
+  - name: Applied Health Analytics for Delivery and Innovation, Nairobi, Kenya
+    index: 2
+  - name: WorldPop, School of Geography and Environmental Science, University of Southampton, Southampton, United Kingdom
+    index: 3
+  - name: Department of Statistics, Nnamdi Azikiwe University, Awka, Nigeria
+    index: 4
+date: 8 July 2025
+bibliography: references.bib
+---
+
+## Summary
+
+`AgePopDenom` is an R package that generates spatial resolution (5 km × 5 km) estimates of population by single year of age across space using publicly available datasets including Demographic and Health Surveys (DHS) [@fabic2012], WorldPop gridded population data [@tatem2017worldpop], and administrative boundaries. It’s designed for settings where recent, detailed age data are missing or outdated, common in many low- and middle-income countries.
+
+The package models age distributions using a Gamma distribution framework, which is well-suited for representing the typical shape of age structures (e.g., more children than elderly in most settings). It then predicts the distribution parameters across space using bivariate geostatistical models, which allow the model to estimate how age patterns vary jointly with location. To better reflect real-world patterns, the model incorporates spatial autocorrelation via Gaussian processes, which is a way of capturing the phenomenon where nearby locations tend to have similar demographic characteristics. Users can also include covariates, such as urban/rural status or other geographic features, to guide predictions. This approach moves beyond standard demographic methods by modeling age structure as a high-resolution surface across space, capturing spatial heterogeneity rather than assuming uniformity within large regions. As a result, AgePopDenom can produce detailed, spatially continuous age-disaggregated estimates, even in locations not directly sampled in surveys.
+
+`AgePopDenom` is, to our knowledge, the first `R` package to operationalize a parameter-based geostatistical approach for modeling full age structures at fine spatial resolution (Figure 1). Unlike existing tools that focus on demographic simulation or a limited set of health indicators, `AgePopDenom` integrates survey microdata, gridded population rasters, administrative boundaries, and user-defined covariates into a single, streamlined workflow. It automates every step, from data acquisition and preprocessing to model fitting, spatial prediction, and visualization. The spatial modeling of age parameters is implemented using Template Model Builder (TMB) [@kristensen2016], an `R` package designed for fitting statistical latent variable models with efficient `C++`-based optimization. This approach reduces computation time to just 7–10 minutes per country, compared to the 2–3 hours typically required by conventional `R` methods.
+
+![Figure 1:AgePopDenom workflow schematic showing the integration of input datasets, core processing pipeline, and resulting outputs.](schematic_figure_final.png)
+
+## Statement of need
+
[…]
