+++ README.md
@@ -1,256 +1,43 @@
-# AgePopDenom <img src="man/figures/logo.png" align="right" height="138"/>
+# AgePopDenom: Geostatistical Modeling for Fine-Scale Population Age-Structures
 
-<!-- badges: start -->
-[![R-CMD-check](https://github.com/truenomad/AgePopDenom/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/truenomad/AgePopDenom/actions/workflows/R-CMD-check.yaml) [![CodeFactor](https://www.codefactor.io/repository/github/truenomad/agepopdenom/badge)](https://www.codefactor.io/repository/github/truenomad/agepopdenom) [![codecov](https://codecov.io/gh/truenomad/AgePopDenom/graph/badge.svg?token=UL9XKIIXTQ)](https://app.codecov.io/gh/truenomad/AgePopDenom) 
-<!-- badges: end -->
+<!-- [![JOSS Status](https://joss.theoj.org/papers/10.21105/joss.xxxxx/status.svg)](https://joss.theoj.org/papers/10.21105/joss.xxxxx) -->
 
-## What is AgePopDenom?
+This repository contains the JOSS paper submission for the **AgePopDenom** R package, which generates spatial resolution (5 km × 5 km) estimates of population by single year of age across space using publicly available datasets.
 
-**`AgePopDenom`** is an R package designed to facilitate the generation of fine-scale, age-structured population denominators for public health decision-making and service delivery. By combining census and household survey data with a novel parameter-based geostatistical modeling approach, the package produces high-resolution (5km x 5km) population estimates disaggregated by age.
+## About AgePopDenom
 
-------------------------------------------------------------------------
+`AgePopDenom` is the first R package to operationalize a parameter-based geostatistical approach for modeling full age structures at fine spatial resolution. Unlike existing tools that focus on demographic simulation or limited health indicators, AgePopDenom integrates:
 
-## Installation
+- Survey microdata (DHS, MICS, or custom household surveys)
+- WorldPop gridded population rasters
+- Administrative boundaries
+- User-defined covariates (urban/rural status, etc.)
 
-### System Requirements
+The package models age distributions using a Gamma distribution framework with bivariate Gaussian processes to capture spatial autocorrelation, delivering high-resolution estimates with built-in uncertainty quantification.
 
-Before installing **AgePopDenom**, ensure your system meets the following requirements:
+## Repository Contents
 
-1.  **R version**: \>= 4.1.0
-2.  **C++ compiler**: C++17 compatible
-3.  **TMB** (Template Model Builder)
+- `paper.md` - Main JOSS manuscript file
+- `references.bib` - Bibliography
+- `paper.html` - Rendered HTML version
+- `schematic_figure_final.png/.html` - Workflow diagram
+- **`example_pipeline.R`** - Complete example demonstrating package functionality
+- **`renv.lock`** - renv lockfile with package dependencies
+- **`AgePopDenom_Paper_JOSS.Rproj`** - RStudio Project file
 
-#### Platform-Specific Setup
+## Getting Started
 
-**Windows**
+### Prerequisites
 
-1.  Install Rtools (matches your R version):
+Before running the example, ensure your system is properly configured:
 
-``` r
-# Check if Rtools is installed and properly configured
[…]
