### 2026-01-15 #8881 @truenomad
@arfon Ahh got ya! This has been sorted now. The paper is in the repo under the paper branch - https://github.com/truenomad/AgePopDenom/tree/paper
