### 2026-01-16 85cf23e
Update JOSS paper to meet new submission requirements
