+++ paper.md
@@ -3,6 +3,7 @@ title: "AgePopDenom: Geostatistical Modeling for Fine-Scale Population Age-Struc
 tags:
   - R
   - geostatistics
+  - geospatial
   - population modeling
   - demographics
   - spatial analysis
@@ -34,52 +35,50 @@ date: 8 July 2025
 bibliography: references.bib
 ---
 
-## Summary
+# Summary
 
-`AgePopDenom` is an R package that generates spatial resolution (5 km × 5 km) estimates of population by single year of age across space using publicly available datasets including Demographic and Health Surveys (DHS) [@fabic2012], WorldPop gridded population data [@tatem2017worldpop], and administrative boundaries. It’s designed for settings where recent, detailed age data are missing or outdated, common in many low- and middle-income countries.
+`AgePopDenom` is an R package that generates high-resolution (5 km × 5 km) estimates of population by single year of age using publicly available datasets including Demographic and Health Surveys (DHS) [@fabic2012], WorldPop gridded population data [@tatem2017worldpop], and administrative boundaries. It is designed for settings where recent, detailed age data are missing or outdated, common in many low- and middle-income countries.
 
-The package models age distributions using a Gamma distribution framework, which is well-suited for representing the typical shape of age structures (e.g., more children than elderly in most settings). It then predicts the distribution parameters across space using bivariate geostatistical models, which allow the model to estimate how age patterns vary jointly with location. To better reflect real-world patterns, the model incorporates spatial autocorrelation via Gaussian processes, which is a way of capturing the phenomenon where nearby locations tend to have similar demographic characteristics. Users can also include covariates, such as urban/rural status or other geographic features, to guide predictions. This approach moves beyond standard demographic methods by modeling age structure as a high-resolution surface across space, capturing spatial heterogeneity rather than assuming uniformity within large regions. As a result, AgePopDenom can produce detailed, spatially continuous age-disaggregated estimates, even in locations not directly sampled in surveys.
+The package models age distributions using a Gamma distribution framework, well-suited for representing typical age structures. It predicts distribution parameters across space using bivariate geostatistical models with Gaussian processes, capturing spatial autocorrelation in demographic characteristics. Users can include covariates such as urban/rural status to guide predictions. This approach models age structure as a high-resolution surface, capturing spatial heterogeneity rather than assuming uniformity within large regions, producing spatially continuous age-disaggregated estimates even in unsampled locations (\autoref{fig:workflow}).
 
-`AgePopDenom` is, to our knowledge, the first `R` package to operationalize a parameter-based geostatistical approach for modeling full age structures at fine spatial resolution (Figure 1). Unlike existing tools that focus on demographic simulation or a limited set of health indicators, `AgePopDenom` integrates survey microdata, gridded population rasters, administrative boundaries, and user-defined covariates into a single, streamlined workflow. It automates every step, from data acquisition and preprocessing to model fitting, spatial prediction, and visualization. The spatial modeling of age parameters is implemented using Template Model Builder (TMB) [@kristensen2016], an `R` package designed for fitting statistical latent variable models with efficient `C++`-based optimization. This approach reduces computation time to just 7–10 minutes per country, compared to the 2–3 hours typically required by conventional `R` methods.
[…]
