### 2026-01-28 #8881 @truenomad
@arfon thanks! Happy to add this. Can you remind me what the word count is now?
