### 2026-01-29 4c8eeb1
Add State of the field section
