+++ paper.md
@@ -51,13 +51,17 @@ Accurate and spatially explicit age-structured population data are essential for
 
 Despite advances in fine-scale population modeling, particularly for total population and select age groups like under-fives [@stevens2015; @wardrop2018; @alegana2015; @pezzulo2017], methods for estimating full age distributions remain limited. Traditional techniques like spline interpolation [@mcneil1977; @fukuda2010] lack spatial detail, top-down census disaggregation relies on outdated baselines, and bottom-up models typically exclude age [@wardrop2018]. AgePopDenom addresses this gap by modeling continuous age distributions via Gamma parameterization with bivariate Gaussian processes, delivering high-resolution estimates with built-in uncertainty quantification.
 
+# State of the field
+
+Several existing R packages address adjacent topics but none support empirical, geostatistical modeling of full age structures at fine spatial resolution. Simulation-oriented tools such as `IBMPopSim` [@giorgi2023efficient], `demogR` [@demogR], `demography` [@demography2024], `popdemo` [@stott2012popdemo], and `mpmsim` [@jones2025mpmsim] focus on cohort or matrix-based demographic modeling rather than empirical spatial estimation. `SUMMER` [@summer2025] supports small-area estimation of survey-based indicators but is limited to specific outcomes like under-five mortality. Similarly, `ungroup` [@pascariu2018ungroup], `sptotal` [@higham2019sptotal], and `APCtools` [@bauer2018apctools] address age unbinning, spatial aggregation, and cohort analysis respectively, without modeling full age structures spatially. The `wpp2024` [@wpp2024] package, based on official UN model life table projections, provides standardized national age patterns but lacks spatial granularity and the ability to incorporate covariates.
+
+`AgePopDenom` fills this gap by operationalizing a parameter-based geostatistical approach for modeling full age structures. Rather than contributing to existing packages, we built a standalone tool because the core requirement—joint spatial prediction of Gamma distribution parameters via bivariate Gaussian processes—is architecturally distinct from the single-outcome focus of packages like `SUMMER` or the simulation paradigms of demographic modeling tools. `AgePopDenom` integrates well-maintained packages (`rdhs` for survey microdata, `terra` for gridded population rasters) into a reproducible pipeline with modular functions for each workflow stage.
+
 # Software design
 
 `AgePopDenom`'s architecture reflects three key design decisions: (1) a parameter-based approach to age modeling, (2) computational efficiency through compiled code, and (3) an end-to-end automated workflow.
 
 Rather than storing and interpolating raw age counts, which would be memory-intensive and noisy, we model age distributions via Gamma parameters (shape and scale), reducing the problem to predicting two smooth spatial surfaces. This parameterization naturally captures typical age-structure shapes while enabling uncertainty quantification through the fitted distribution. For computational efficiency, spatial modeling is implemented using Template Model Builder (TMB) [@kristensen2016], which compiles model specifications to C++ and uses automatic differentiation for optimization, making the approach practical for operational use.
-
[…]
