### 2026-02-17 #8881 @strengejacke
Sure, happy to review this paper!
