### 2026-03-04 #8881 @fangzhou-xie
@ConnorDonegan, would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
