### 2026-03-27 #8881 @garretrc
> [@garretrc](https://github.com/garretrc), would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

Yes, I'm willing to review!
