### 2026-03-27 #8881 @mponce0
> [@mponce0](https://github.com/mponce0), would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

Sorry, I'm not sure whether you are still looking for reviewers.
If that's the case, I'm happy to help but not until the end of the academic term in a couple of weeks.
Please let me know if that would work for you.
