### 2026-03-31 #10335 @editorialbot
[REVIEW]: AgePopDenom: Geostatistical Modeling for Fine-Scale Population Age-Structures
**Submitting author:** @truenomad (<a href="http://orcid.org/0000-0002-9339-4613">Mohamed Yusuf</a>)
**Repository:** https://github.com/truenomad/AgePopDenom
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.2.3
