### 2026-04-03 #10335 @fangzhou-xie
Hi @truenomad, thanks for your submission to JOSS. We have the two reviewers for this paper, and will start the review soon. In the meantime, please fix the "doi" issue as listed by the editorial bot above.
