### 2026-04-13 #10335 @garretrc
- [x] **Open development:** (Note: I recommend adding a github release after joss but the CRAN release obviously satisfies this requirement) Was the software developed openly from early stages? For projects with recent public repositories, is there at least six months of public development history with evidence of releases, public issues/pull requests, and ideally external engagement?
