### 2026-05-18 truenomad/AgePopDenom#3 @garretrc
Tiny improvements for the paper
I recognize these are small nitpicks, the paper generally looks great! [openjournals/joss-reviews#10335](https://github.com/openjournals/joss-reviews/issues/10335)

Figure 1: text is a bit small for reading on low res screens, quick option that might help a bit is to remove some of the white margin in the image.

Line 110: code runs over the margin a bit.

Feel free to close once addressed.
