### 2026-05-24 2788eee
Address reviewer feedback: wrap long code line and trim figure margins
