### 2026-05-24 37542e1
Add DHS data access section to README and fix template typos
