+++ articles/index.md
@@ -1,0 +1,6 @@
+# Articles
+
+### Using AgePopDenom
+
+- [Using
+  AgePopDenom](https://truenomad.github.io/AgePopDenom/articles/using_agepopdenom.md):
