+++ LICENSE-text.md
@@ -1,0 +1,4 @@
+# License
+
+    YEAR: 2025
+    COPYRIGHT HOLDER: AgePopDenom authors

[…]
