### 2026-05-24 b27be98
Add missing DOIs and fix sptotal citation
