+++ paper.md
@@ -53,7 +53,7 @@ Despite advances in fine-scale population modeling, particularly for total popul
 
 # State of the field
 
-Several existing R packages address adjacent topics but none support empirical, geostatistical modeling of full age structures at fine spatial resolution. Simulation-oriented tools such as `IBMPopSim` [@giorgi2023efficient], `demogR` [@demogR], `demography` [@demography2024], `popdemo` [@stott2012popdemo], and `mpmsim` [@jones2025mpmsim] focus on cohort or matrix-based demographic modeling rather than empirical spatial estimation. `SUMMER` [@summer2025] supports small-area estimation of survey-based indicators but is limited to specific outcomes like under-five mortality. Similarly, `ungroup` [@pascariu2018ungroup], `sptotal` [@higham2019sptotal], and `APCtools` [@bauer2018apctools] address age unbinning, spatial aggregation, and cohort analysis respectively, without modeling full age structures spatially. The `wpp2024` [@wpp2024] package, based on official UN model life table projections, provides standardized national age patterns but lacks spatial granularity and the ability to incorporate covariates.
+Several existing R packages address adjacent topics but none support empirical, geostatistical modeling of full age structures at fine spatial resolution. Simulation-oriented tools such as `IBMPopSim` [@giorgi2023efficient], `demogR` [@demogR], `demography` [@demography2024], `popdemo` [@stott2012popdemo], and `mpmsim` [@jones2025mpmsim] focus on cohort or matrix-based demographic modeling rather than empirical spatial estimation. `SUMMER` [@summer2025] supports small-area estimation of survey-based indicators but is limited to specific outcomes like under-five mortality. Similarly, `ungroup` [@pascariu2018ungroup], `sptotal` [@higham2023sptotal], and `APCtools` [@bauer2018apctools] address age unbinning, spatial aggregation, and cohort analysis respectively, without modeling full age structures spatially. The `wpp2024` [@wpp2024] package, based on official UN model life table projections, provides standardized national age patterns but lacks spatial granularity and the ability to incorporate covariates.
 
 `AgePopDenom` fills this gap by operationalizing a parameter-based geostatistical approach for modeling full age structures. Rather than contributing to existing packages, we built a standalone tool because the core requirement—joint spatial prediction of Gamma distribution parameters via bivariate Gaussian processes—is architecturally distinct from the single-outcome focus of packages like `SUMMER` or the simulation paradigms of demographic modeling tools. `AgePopDenom` integrates well-maintained packages (`rdhs` for survey microdata, `terra` for gridded population rasters) into a reproducible pipeline with modular functions for each workflow stage.
 
