+++ references.bib
@@ -77,7 +77,8 @@
   volume={12},
   pages={1--11},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/1471-2458-12-741}
 }
 
 @article{kristensen2016,
@@ -106,7 +107,8 @@
   volume={90},
   pages={604--612},
   year={2012},
-  publisher={SciELO Public Health}
+  publisher={SciELO Public Health},
+  doi={10.2471/blt.11.095513}
 }
 
 @article{yankey2025,
@@ -124,7 +126,8 @@
   number={14},
   pages={3529--3537},
   year={2018},
-  publisher={National Academy of Sciences}
+  publisher={National Academy of Sciences},
+  doi={10.1073/pnas.1715305115}
 }
 
 @article{pezzulo2017,
@@ -135,7 +138,8 @@
   number={1},
   pages={1--15},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/sdata.2017.89}
 }
 
 @article{james2018,
@@ -146,7 +150,8 @@
   number={1},
   pages={1--11},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/sdata.2018.90}
 }
 
 @article{alegana2015,
@@ -157,7 +162,8 @@
   number={105},
   pages={20150073},
   year={2015},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  doi={10.1098/rsif.2015.0073}
 }
 
 @article{mcneil1977,
@@ -177,7 +183,8 @@
   volume={27},
   pages={1--13},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12546-010-9028-z}
 }
 
 @article{watson2019,
@@ -217,7 +224,8 @@
   number={1},
   pages={127--139},
   year={2005},
-  publisher={SAGE Publications Sage UK: London, England}
+  publisher={SAGE Publications Sage UK: London, England},
+  doi={10.1068/a36202}
 }
 
 
@@ -266,7 +274,8 @@
   journal={Methods in Ecology \& Evolution},
   volume={3},
   number={5},
-  year={2012}
+  year={2012},
+  doi={10.1111/j.2041-210x.2012.00222.x}
 }
 
 @article{jones2025mpmsim,
@@ -293,6 +302,7 @@
   year = {2024},
   note = {R package version 2.0.0.9000, https://github.com/robjhyndman/demography},
   url = {https://pkg.robjhyndman.com/demography/},
+  doi = {10.32614/cran.package.demography}
 }
 
 @article{rue2017,
@@ -324,7 +334,8 @@
   number={1},
   pages={1--4},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/sdata.2017.4}
 }
 
 @article{pascariu2018ungroup,
@@ -340,16 +351,16 @@
   langid  = {en}
 }
 
-@article{higham2019sptotal,
-  title   = {sptotal: An {R} package for predicting totals and weighted sums from spatial data},
-  author  = {Higham, Matthew and de Souza, Leandro},
+@article{higham2023sptotal,
+  title   = {sptotal: an {R} package for predicting totals and weighted sums from spatial data},
+  author  = {Higham, Matt and Ver Hoef, Jay and Frank, Bryce and Dumelle, Michael},
   journal = {Journal of Open Source Software},
-  volume  = {4},
-  number  = {42},
-  pages   = {1679},
-  year    = {2019},
[…]
