### 2026-07-05 0f8f80e
Address final JOSS review feedback
- Auto-size future.globals.maxSize from the prediction matrices and
  restore the user's future plan and options on exit, so large countries
  (e.g. Uganda) no longer require manual option changes
- Set up parallel workers once per run instead of per age interval
- Guard n_cores defaults with max(1, detectCores() - 2) for low-core
  machines
- Remove dead run_parallel_mac/run_parallel_other helpers
- Ship a pre-simulated example dataset (inst/extdata/example_dhs_pr_gmb.rds)
- Add a no-DHS-account quick start and example outputs (figures + table)
  to the READ
[…]
