+++ vignettes/using_agepopdenom.Rmd
@@ -145,6 +145,8 @@ You can download or place your own survey data into `01_data/1a_survey_data/proc
 -   *age_in_years* for each individual
 -   An *urban* indicator (if available)
 
+If you do not (yet) have a DHS account, you can still run the entire workflow with simulated data: `simulate_dummy_dhs_pr()` writes a DHS-like dataset to the expected location, and a pre-simulated copy is shipped with the package (`system.file("extdata", "example_dhs_pr_gmb.rds", package = "AgePopDenom")`). See the *Example: Gambia* section below for a complete runnable script.
+
 To download DHS data, do:
 
 ```{r eval=FALSE, include=TRUE}
@@ -606,7 +608,7 @@ AgePopDenom::download_shapefile(cntry_code)
 # download population rasters from worldpop
 AgePopDenom::download_pop_rasters(cntry_code)
 
-# wxtract urban extent raster
+# extract urban extent raster
 AgePopDenom::extract_afurextent()
 
 # Run models and get outputs ---------------------------------------------------
