+++ DESCRIPTION
@@ -65,7 +65,7 @@ Suggests:
     matrixStats,
     testthat,
     RcppEigen
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 VignetteBuilder: knitr
 Config/testthat/edition: 3
 URL: https://truenomad.github.io/AgePopDenom/, https://github.com/truenomad/AgePopDenom

[…]
