### 2026-07-05 89e8585
Add example outputs figure and simulated-data path to Example section
Responds to final review feedback requesting example data and example
output in the paper: the Example section now points to
simulate_dummy_dhs_pr() for a no-credentials run and shows the predicted
Gamma parameter surfaces and mean-age prediction it produces.
