### 2026-07-05 ce73d2c
Add test-coverage workflow to fix 0% Codecov reporting
The Codecov badge showed 0% because CI never generated a coverage
report: R-CMD-check.yaml called codecov/codecov-action (which only
uploads an existing report) without any covr step producing one.

- Add .github/workflows/test-coverage.yaml running covr::package_coverage()
  -> covr::to_cobertura(), then uploading via codecov-action with the
  existing CODECOV_TOKEN secret. NOT_CRAN=true so skip_on_cran() tests run.
- Remove the no-op codecov upload step from R-CMD-check.yaml
- Add covr to Suggests
