+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   using_agepopdenom: using_agepopdenom.html
-last_built: 2026-07-05T10:23Z
+last_built: 2026-07-05T10:31Z
 urls:
   reference: https://truenomad.github.io/AgePopDenom/reference
   article: https://truenomad.github.io/AgePopDenom/articles

[…]
