+++ reference/get_current_script_path.md
@@ -21,5 +21,5 @@ if it cannot be determined.
 
 ``` r
 get_current_script_path()
-#> [1] "/home/runner/work/_temp/4eef36fe-f9ff-4910-a12a-25d506cb9d4b"
+#> [1] "/home/runner/work/_temp/6d3c2f71-0f15-4201-a953-4009ef7cc920"
 ```
