### 2026-07-05 7a10d6a
Fix Codecov upload: use files input for codecov-action v5
codecov/codecov-action@v5 renamed the single-file input from 'file' to
'files'. With 'file' ignored and disable_search on, no report was
uploaded ('No coverage reports found'). Local covr run reports 78.8%.
