+++ reference/generate_age_pop_table.md
@@ -115,6 +115,6 @@ output <- generate_age_pop_table(
 #> ℹ Processing interval 10/10...
 #> ✔ Completed interval 10/10.
 #> 
-#> ✔ Final age population data saved to /tmp/RtmpqFflLL/ABC_age_tables_pop_0_99plus_yrs_by_10yrs.rds
+#> ✔ Final age population data saved to /tmp/Rtmplnf4Dh/ABC_age_tables_pop_0_99plus_yrs_by_10yrs.rds
 # }
 ```
