### 2026-07-05 e6cb743
Use implicit returns and fix a message typo (lint cleanup)
Convert three terminal explicit return() calls to implicit returns
(return_linter) in extract_afurextent(), generate_age_pop_table(), and
process_final_population_data(), and fix 'paramters' -> 'parameters' in a
success message. R source now passes CodeFactor's lintr config with zero
issues.
