+++ reference/process_dhs_data.md
@@ -149,6 +149,6 @@ process_dhs_data(
 #> ℹ Aggregating and extracting gamma for: TZA
 #> ✔ Aggregated and extracted gamma for: TZA (2 of 2)
 #> 
-#> ✔ All countries processed. Combined data saved to /tmp/Rtmplnf4Dh/test_env/output.rds
+#> ✔ All countries processed. Combined data saved to /tmp/Rtmpkz8QDq/test_env/output.rds
 # }
 ```
