### 2026-09-01 #10335 @truenomad
Hi all,

Following up on the AgePopDenom review. I’d be grateful if we could bring the review to a close. Please let me know if there are any outstanding issues or changes needed from my side.

I’m happy to address these promptly. Thanks.
