### 2026-09-14 26fd50e
Update maintainer affiliation to WHO EMRO in README
