### 2026-09-14 39abd01
Update affiliations and fix APCtools, TMB and mpmsim references for JOSS review
