+++ references.bib
@@ -86,8 +86,10 @@
   author={Kristensen, Kasper and Nielsen, Anders and Berg, Casper W and Skaug, Hans and Bell, Bradley M},
   journal={Journal of statistical software},
   volume={70},
+  number={5},
   pages={1--21},
-  year={2016}
+  year={2016},
+  doi={10.18637/jss.v070.i05}
 }
 
 
@@ -279,14 +281,15 @@
 }
 
 @article{jones2025mpmsim,
-  title={mpmsim: An R package for simulating matrix population models},
+  title={{mpmsim}: An {R} package for simulating matrix population models},
   author={Jones, Owen R},
   journal={Methods in Ecology and Evolution},
   volume={16},
   number={5},
   pages={904--911},
   year={2025},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/2041-210X.70030}
 }
 
 @Manual{summer2025,
@@ -364,15 +367,15 @@
   langid  = {en}
 }
 
-@article{bauer2018apctools,
+@article{bauer2022apctools,
   title   = {{APCtools}: Descriptive and Model-Based Age-Period-Cohort Analysis},
-  author  = {Bauer, Alexander},
+  author  = {Bauer, Alexander and Weigert, Maximilian and Jalal, Hawre},
   journal = {Journal of Open Source Software},
-  volume  = {3},
-  number  = {22},
-  pages   = {480},
-  year    = {2018},
-  doi     = {10.21105/joss.00480},
-  url     = {https://doi.org/10.21105/joss.00480},
+  volume  = {7},
+  number  = {73},
+  pages   = {4056},
+  year    = {2022},
+  doi     = {10.21105/joss.04056},
+  url     = {https://doi.org/10.21105/joss.04056},
   langid  = {en}
 }
