### 2026-09-14 600b495
Add city and country to WHO AFRO affiliation
