+++ reference/extract_age_param.md
@@ -52,6 +52,6 @@ saveRDS(dummy_params, file = file.path(dummy_dir,
 
 params_df <- extract_age_param(dir_path = dummy_dir,
                                output_file = tempdir())
-#> ✔ Model parameters extracted and saved to /tmp/Rtmpfnsc2N.
+#> ✔ Model parameters extracted and saved to /tmp/Rtmp5XisFS.
 # }
 ```
