+++ reference/generate_age_pop_raster.md
@@ -131,6 +131,6 @@ res <- generate_age_pop_raster(predictor_data,
 #> ℹ Processing interval 11/11...
 #> ✔ Completed interval 11/11.
 #> 
-#> ✔ Raster stack saved to /tmp/Rtmp5XisFS/cod_age_pop_grid_0_10_yrs_by_1yrs.tif
+#> ✔ Raster stack saved to /tmp/RtmpeF8RAC/cod_age_pop_grid_0_10_yrs_by_1yrs.tif
 # }
 ```
