### 2026-09-14 84ab979
Remove duplicate mcneil1977 bib entry
