+++ references.bib
@@ -319,16 +319,6 @@
   publisher={Annual Reviews}
 }
 
-@article{mcneil1977,
-  title={Spline interpolation of demographic oata},
-  author={McNeil, Donald R and Trullell, T Jamel and Turner, John C},
-  journal={Demography},
-  volume={14},
-  pages={245--252},
-  year={1977},
-  publisher={Springer}
-}
-
 @article{tatem2017worldpop,
   title={WorldPop, open data for spatial demography},
   author={Tatem, Andrew J},
